tests/requirements_tests.txt
pytest
click
requests
psycopg
python-dotenv